requirements.txt
# Core scraping
httpx>=0.27.0
trafilatura>=1.12.0
readability-lxml>=0.8.0
markdownify>=0.13.0
lxml>=5.3.0
selectolax>=0.3.0
rich>=13.8.0
tldextract>=5.1.0
urlextract>=1.9.0
beautifulsoup4>=4.12.0
feedparser>=6.0.0

# QA/Agent
langchain>=0.3.0
langchain-community>=0.3.0
langchain-anthropic>=0.3.0
faiss-cpu>=1.8.0
tiktoken>=0.8.0
numpy<2
python-dotenv>=1.0.0
sentence-transformers>=3.0.0
transformers>=4.0.0
torch>=2.0.0
Pillow>=10.0.0
scikit-learn>=1.0.0
scipy>=1.0.0

# Frontend
streamlit>=1.39.0